chore: refactor folder structures; add new necessary migrations
