+++ app/Models/Org/Company.php
@@ -1,31 +1,29 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Org;
 
 use Illuminate\Database\Eloquent\Model;
-use App\Models\Employee;
-use App\Models\Department;
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Position extends Model
+class Company extends Model
 {
 
-    protected $table = 'tb_position';
+    protected $table = "tb_company";
     protected $primaryKey = 'id';
     public $incrementing = false;
 
-    use HasFactory, Notifiable, SoftDeletes;
+    use HasFactory, Notifiable, HasUuids, SoftDeletes;
+
     /**
      * The attributes that are mass assignable.
      *
      * @var list<string>
      */
     protected $fillable = [
-        'name',
-        'level',
+        'name'
     ];
 
     /**
@@ -34,16 +32,20 @@ class Position extends Model
      * @var list<string>
      */
     protected $hidden = [
-        'id_department',
+        'id_manager',
+        'id_subscription',
+        'effective_date',
+        'created_at',
+        'updated_at',
     ];
+
+    public function manager()
+    {
+        return $this->belongsTo(User::class, 'id_manager');
+    }
 
     public function employees()
     {
-        return $this->hasMany(Employee::class, 'id_position');
-    }
-
-    public function department()
-    {
-        return $this->belongsTo(Department::class, 'id_department');
+        return $this->hasMany(User::class, 'id_company');
     }
 }
