+++ routes/api/auth.route.php
@@ -2,7 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 
-use App\Http\Controllers\AuthController;
+use App\Http\Controllers\Auth\AuthController;
 
 Route::group([
     'prefix' => 'auth',
